include search filters for json and tsv download
